Debug jax tests

+++ pennylane/ops/qubit/state_preparation.py
@@ -628,7 +628,7 @@ def _stateprep_resources(num_wires):
     return {
         subroutine_resource_rep(
             qml.MottonenStatePreparation,
-            AbstractArray((2**num_wires,), dtype=float),
+            AbstractArray((2**num_wires,)),
             AbstractArray((num_wires,)),
         ): 1
     }
